Added the possibility to add a custom TF-M config
TF-M offers the possibility to add a custom config file via an input argument to be parse in priority during the build. The issue is that Zephyr doesn't give any way to pass this argument.

This commit adds a way to pass that argument automatically when a tfm/config.cmake file is present in the project folder.

Signed-off-by: Farès Chati <fares.chati@intellinium.com>